Expand shared events
