+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_doughnut.ex
@@ -4,30 +4,14 @@ defmodule UnchartedPhoenix.LiveDoughnutComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def mount(socket) do
-    {:ok, assign(socket, :show_table, false)}
-  end
+  use UnchartedPhoenix.SharedEvents
 
   def update(assigns, socket) do
     socket =
       socket
-      |> assign(:chart, assigns.chart)
+      |> shared_update(assigns)
       |> assign(:doughnut_slices, Uncharted.DoughnutChart.doughnut_slices(assigns.chart))
-      |> assign(:always_show_table, assigns.always_show_table)
 
     {:ok, socket}
-  end
-
-  def render(assigns) do
-    Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_doughnut.html", assigns)
-  end
-
-  def handle_event("show_table", _, socket) do
-    {:noreply, assign(socket, :show_table, true)}
-  end
-
-  def handle_event("hide_table", _, socket) do
-    {:noreply, assign(socket, :show_table, false)}
   end
 end
