+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_horizontal_funnel.ex
@@ -4,10 +4,7 @@ defmodule UnchartedPhoenix.LiveHorizontalFunnelComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def mount(socket) do
-    {:ok, assign(socket, :show_table, false)}
-  end
+  use UnchartedPhoenix.SharedEvents
 
   def update(assigns, socket) do
     x_axis = assigns.chart.dataset.axes.magnitude_axis
@@ -22,12 +19,11 @@ defmodule UnchartedPhoenix.LiveHorizontalFunnelComponent do
 
     socket =
       socket
-      |> assign(:chart, assigns.chart)
+      |> shared_update(assigns)
       |> assign(:columns, Uncharted.HorizontalFunnelChart.bars(assigns.chart))
       |> assign(:grid_lines, grid_lines)
       |> assign(:offsetter, grid_line_offsetter)
       |> assign(:axis, x_axis)
-      |> assign(:always_show_table, assigns.always_show_table)
       |> assign(:width, assigns.chart.width || 600)
       |> assign(:height, assigns.chart.height || 400)
 
@@ -36,13 +32,5 @@ defmodule UnchartedPhoenix.LiveHorizontalFunnelComponent do
 
   def render(assigns) do
     Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_horiz_funnel.html", assigns)
-  end
-
-  def handle_event("show_table", _, socket) do
-    {:noreply, assign(socket, :show_table, true)}
-  end
-
-  def handle_event("hide_table", _, socket) do
-    {:noreply, assign(socket, :show_table, false)}
   end
 end
