+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_pie.ex
@@ -4,19 +4,14 @@ defmodule UnchartedPhoenix.LivePieComponent do
   """
 
   use Phoenix.LiveComponent
-  use UnchartedPhoenix.TableEvents
+  use UnchartedPhoenix.SharedEvents
 
   def update(assigns, socket) do
     socket =
       socket
-      |> assign(:chart, assigns.chart)
+      |> shared_update(assigns)
       |> assign(:pie_slices, Uncharted.PieChart.pie_slices(assigns.chart))
-      |> assign(:always_show_table, assigns.always_show_table)
 
     {:ok, socket}
-  end
-
-  def render(assigns) do
-    Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_pie.html", assigns)
   end
 end
